+++ src/main/java/io/github/sakaki_aruka/customcrafter/api/interfaces/recipe/CRecipe.kt
@@ -21,6 +21,34 @@ interface CRecipe {
     val results: List<ResultSupplier>?
     val filters: Set<CRecipeFilter<CMatter>>?
     val type: CRecipeType
+
+    companion object {
+        /**
+         * Returns a specified [CRecipe] is valid or not.
+         * ```kotlin
+         * val recipe: CRecipe = ~~~
+         * CRecipe.isValidCRecipe(recipe).exceptionOrNull()?.let { throw it }
+         * ```
+         * @return[Result] Result of check
+         * @since v5.0.14
+         */
+        fun isValidCRecipe(recipe: CRecipe): Result<Unit> {
+            return if (recipe.items.isEmpty() || recipe.items.size > 36) {
+                Result.failure(IllegalStateException("'items' must contain 1 to 36 valid CMatters."))
+            } else if (recipe.items.values.any { matter -> CMatter.isValidCMatter(matter).isFailure }) {
+                val builder = StringBuilder()
+                for ((c, matter) in recipe.items.entries) {
+                    val t: Throwable = CMatter.isValidCMatter(matter).exceptionOrNull()
+                        ?: continue
+                    builder.append("[items] x: ${c.x}, y: ${c.y}, ${t.message} ${System.lineSeparator()}")
+                }
+                Result.failure(IllegalStateException(builder.toString()))
+            } else {
+                Result.success(Unit)
+            }
+        }
+    }
+
     /**
      * replace [items]
      *